Seperated greeting to greeting.js
